requirements.txt
openai
google-generativeai
huggingface_hub
python-dotenv
numpy
requests
